client_servidor/requirements.txt
fastapi
uvicorn
paho-mqtt==1.6.1
jinja2
python-dotenv 